Make unit test working with PHP8.3

+++ test/Horde/Argv/ExtendAddTypesTest.php
@@ -14,12 +14,14 @@ use \Horde_Argv_Option;
 
 class ExtendAddTypesTest extends TestCase
 {
+    private string $testPath;
+
     public function setUp(): void
     {
         parent::setUp();
-        if (class_exists('Horde_Argv_ExtendAddTypesTest_MyOption')) {
+        if (class_exists('Horde\\Argv\\ExtendAddTypesTest\\MyOption')) {
             $this->parser = new InterceptingParser(array('usage' => Horde_Argv_Option::SUPPRESS_USAGE,
-                                                                    'optionClass' => 'Horde_Argv_ExtendAddTypesTest_MyOption'));
+                                                                    'optionClass' => 'Horde\\Argv\\ExtendAddTypesTest\\MyOption'));
             $this->parser->addOption("-a", null, array('type' => "string", 'dest' => "a"));
             $this->parser->addOption("-f", "--file", array('type' => "file", 'dest' => "file"));
         }
@@ -39,7 +41,7 @@ class ExtendAddTypesTest extends TestCase
 
     public function testFiletypeOk()
     {
-        if (class_exists('Horde_Argv_ExtendAddTypesTest_MyOption')) {
+        if (class_exists('Horde\\Argv\\ExtendAddTypesTest\\MyOption')) {
             touch($this->testPath);
             $this->assertParseOK(array("--file", $this->testPath, "-afoo"),
                                 array('file' => $this->testPath, 'a' => 'foo'),
@@ -52,17 +54,24 @@ class ExtendAddTypesTest extends TestCase
 
     public function testFiletypeNoexist()
     {
-        unlink($this->testPath);
-        $this->markTestIncomplete();
+        if (class_exists('Horde\\Argv\\ExtendAddTypesTest\\MyOption')) {
+            unlink($this->testPath);
+            $this->assertParseFail(array("--file", $this->testPath, "-afoo"),
+                               sprintf("%s: file does not exist", $this->testPath));
+        } else {
+            $this->markTestSkipped('Class Horde_Argv_ExtendAddTypesTest_MyOption doesnt exist.');
+        }
     }
 
     public function testFiletypeNotfile()
     {
-        $this->expectException('ReflectionException');
-        unlink($this->testPath);
-        mkdir($this->testPath);
-        $this->assertParseFail(array("--file", $this->testPath, "-afoo"),
+        if (class_exists('Horde\\Argv\\ExtendAddTypesTest\\MyOption')) {
+            unlink($this->testPath);
+            mkdir($this->testPath);
+            $this->assertParseFail(array("--file", $this->testPath, "-afoo"),
                                sprintf("%s: not a regular file", $this->testPath));
+        } else {
+            $this->markTestSkipped('Class Horde_Argv_ExtendAddTypesTest_MyOption doesnt exist.');
+        }
     }
-
 }
